Update ND_BYTES_AVAILABLE_AFTER() macro for better accuracy
With this change the number of bytes available in the captured data
given by the macro is 0 when the argument is greater than or equal
to ndo_snapend or less than ndo_packetp (e.g. pointer underflow).

i.e.:
If p is like p3, the macro gives 0.
If p is like p1, the macro gives 0.

|------------------|-----------------------|-----------------|
      p1  <   ndo_packetp  <=  p2  <  ndo_snapend   <=   p3

Update the ascii_print(), hex_and_ascii_print_with_offset() and
hex_print_with_offset() functions accordingly.

This is a follow-up to 07a7f33ba3ff346b9ce31772645a06980baca907.