CI: Set DELETE_PREFIX in build.sh. [skip ci]
When the PREFIX variable is set before build.sh runs, this means it
points to a directory that build.sh needs neither to create nor to
remove, as that's a responsibility of whoever runs build.sh (which is
usually build_matrix.sh).  When the variable is not set before build.sh
runs, it means build.sh needs to create a temporary directory and to
remove it afterwards.

This convention works in libpcap and tcpslice, but in tcpdump the script
does not delete the temporary directory because I did not port the logic
correctly in commit 75e92be.  Get this right.