Make nd_trunc_longjmp() not static inline.
It should rarely be called (if it's called, either the packet was cut
short by snapshot-based slicing or it's malformed, or the dissector has
a bug), so inlining it doesn't help much, and, as it calls longjmp(),
which often expands to a call to __builtin_longjmp(), GCC won't inline
it anyway (see "An Inline Function is As Fast As a Macro" in the GCC
documentation), and Apple clang version 14.0.3 (clang-1403.0.22.14.1),
at least, doesn't inline it, either, so you get a bunch of inlined
definitions of it throughout the generated code.

Just make it a regular routine.