pgm: fix the way we step through the packet.
Step past the PGM header after we finish processing it and before we
process the message-type-specific header.

Step past the message-type-specific fixed-length header before we
process the stuff after that header.

This makes the code a bit clearer (by explicitly advancing bp by the
size of the stuff we just processed, rather than doing so by trickery
involving adding 1 to a pointer to a structure), and fixes the
processing of message types that don't have a message-type-specific
header (where we weren't stepping past the PGM header).  It also affects
the way we handle messages of an unknown type.