ICMPv6: Recognise ND option 38 (PREF64)
This option, defined in RFC 8781, allows a router administrator to pass
NAT64 prefix information for the network to end hosts together with
other routing and prefix information in the RA message, getting rid
of the need for DNS64 service in the network.

As of today the option is widely supported by software routers,
including radvd, BIRD and systemd-networkd.

We implement the printer and test some valid and broken option contents.