Remove old debug code from print-ascii.c. [skip ci]
Since commit ac910c0 in 2014 the main() function inside "#ifdef MAIN"
cannot be compiled because it does not match the prototypes of the
functions it tests.  Remove the conditional block instead of fixing it
because the existing tests among other things cover the functions in the
file.  This addresses GitHub bug #1079.