Retire BSD/OS support.
The last release of BSD/OS was more than 20 years ago, so remove all
bits specific to building tcpdump on it, mainly ppp_bsdos_if_print(),
which has been a no-op on platforms other than BSD/OS.  Restoring the
function would not be practicable because in addition to the generic bit
rot it uses SLC_DIR, SLC_LLHL, SLC_BPFHDRLEN, SLC_CHL and struct
ppp_header, none of which appear in FreeBSD or illumos or Linux or
NetBSD or OpenBSD.  This is why protocol decoders should not depend on
OS-specific headers.

See also commit 4729260.