LISP: Simplify some codes
Avoid using 'packet_iterator + packet_offset' and loc_ip_pointer,
because they have the same value at the time they are used.

Remove the loc_ip_pointer variable and it's initialization.

This change shows better that a ND_TCHECK_4() is useless because
GET_IPADDR_STRING() is used. Remove it.
This change shows better that a ND_TCHECK_16() is useless because
GET_IP6ADDR_STRING() is used. Remove it.