CHANGES: move backported change to 4.99 section, clean up other stuff. [skip ci]
The elimination of configure and config.h.in from the repository and the
addition of the autogen.sh script to generate those using autoconf was
backported to the 4.99 section; move it to that section.

Fix capitalization in the DHCPv6 entry, to match changes made in the
4.99 branch.

Fix the capitalization of "autoconf" in the 5.0.0 and 4.99.5 sections to
consistently not capitalize it.  (If somebody prefers it to be
capitalized, please consistently capitalize it.)