Disregard setlinebuf(3), always use setvbuf(3).
As most setlinebuf(3) man pages mention, setlinebuf() is a shorthand for
setvbuf().  The latter is in C99, but the former is not, so it is most
logical to use setvbuf() in all cases and to lose another build-time
check.

(cherry picked from commit a9c7cf884805b0a3ebdefd574e5a061f31c5e3fb)