man: Refine for better HTML output. [skip ci]
Format man page references, filter expressions, command-line options and
C names consistently with most other man pages.  Fixup capitalization
and wording in a few instances, among other things do not document NFS
as Sun-specific and AFS -- as Transarc-specific.  Remove the reference
to bpf(4), which belongs to libpcap only, and the description of bugs
that are entirely libpcap-specific and are documented in pcap-filter(7).