SMB: clean up date handling.
Pull the routine that carves a DOS date/time up and puts it into a
struct tm into the routine that converts the struct tm into a time_t.

Note a place where a DOS date/time might not fit in a 32-bit integral
time_t (DOS dates and times can go past the Y2.038K date/time).

Do the NT FILETIME to time_t conversion as an all-integer operation.
Catch cases where the result doesn't fit into a time_t.  (And rename the
routine to speak of it as a FILETIME; that's what it's called.)