NTP: Print kiss codes relevant for NTP debugging
NTPv4 (RFC 5905, section 7.4) formalizes that if the stratum value
is zero, the refid _may_ contain a printable ("designed for
character displays and log files"), four-character, left justified,
zero filled ASCII string for status reporting and debugging ("kiss
code"). RFC 5905 establishes a registry at IANA for kiss codes,
initially populated with values defined in the RFC, but the list
may be modified or extended in the future, and unregistered kiss
codes, e.g., for experimentation and development, are possible (and
are being seen in the field).

Thus, IF the stratum value is zero AND the refid field begins with
a printable character, assume the refid field contains a kiss code
and print it (instead of current "(unspec)").