RSVP: Adjust P2MP ID rendering to more common hex format.
The P2MP ID is typically denoted as a single value, either in decimal or in
hexadecimal format. The 4-octet notation is quite uncommon and I believe this
was just a mistake in the code since the IPv6 version 11 lines above uses a
hexadecimal notation instead.

I'm also using this opportunity to change the name from "P2MP LSP ID" to "P2MP
ID" since that's more correct. The term "LSP ID" has a well-defined meaning in
the RSVP protocol and is something else than the P2MP ID. The BGP dissector
uses the term P2MP ID as well for the same value (and shows it in hexadecimal
format).