tests: Add two ARCnet packet captures. [skip appveyor]
The capture files were produced using an improvised ARCnet+Ethernet+LTE
network at Museum Of Communication in Burntisland.

ARCnet node 0xBE (the IPv4 host) is a Debian 3.0 PC with kernel 2.4.18
and a UM9065L-based ARCnet interface only.  ARCnet node 0x50 (the IPv4
router) is a Debian 6.0 PC with kernel 2.6.32, a COM20020-based ARCnet
interface and an Ethernet uplink to an LTE router (the latter also
appears in the captures as the DNS server).

The host pings the router and then makes an HTTP request to a web server
on the Internet, there is some ARP exchange as well.  One capture
represents this being done using RFC 1051 mode and another -- RFC 1201.