Raw IPv4/IPv6: Add ipv4_if_print() and ipv6_if_print() printers
With this change tcpdump prints errors when an IPv6 packet is in a
LINKTYPE_IPV4 file and when an IPv4 packet is in a LINKTYPE_IPV6 file.

Update the printer summary.

And some test files printed with or without '-e'.

Update the linktype of hoobr_rt6_print.pcap from "Raw IPv4" to
"Raw IPv6" to keep test validity for packet #3.
Update 2 test outputs.