icmp6: add RA PIO PD Preferred Flag
This is a 1-bit flag in the Prefix Information RA option defined by
RFC9762. If a compliant RA-accepting node receives a PIO with the P flag
set, it is expected to make use of DHCPv6 Prefix Delegation to obtain a
prefix and assign addresses from that prefix, not individual addresses
from the PIO-specified prefix.

Link: https://tools.ietf.org/html/rfc9762
Link: https://tools.ietf.org/html/rfc9663
Link: https://www.iana.org/assignments/icmpv6-parameters/icmpv6-parameters.xhtml#ipv6-neighbor-discovery-prefix-information-options