tests: Add coverage for GSO/TSO and BIG TCP
This commit adds coverage that affirms the current behavior of tcpdump
when it parses packets captured in Linux with aggregation offloads
enabled (such as GRO, GSO, TSO and BIG TCP).

New tests include:

1. GRO/GSO/TSO TCP packets. These have ip_len (IPv4) or payload_len
(IPv6) set to a value bigger than MTU, but smaller than 64k (fits the
width of the length field, which is 16 bits).

2. BIG TCP packets, as currently implemented in Linux. These aggregated
packets are larger than 64k. In IPv4, ip_len is just set to 0. In IPv6,
payload_len is also set to 0, and the real length of the payload is
indicated in the Jumbo option of the Hop-by-hop extension header,
immediately following the IPv6 header.

The ipv4_tcp_http_xml_tso test got removed. It contained a small (<64k)
IPv4 packet with ip_len set to 0, and tcpdump treated it as valid,
printing a message that TSO was presumed. My tests show that TSO
actually fills in the correct (non-zero) value into the ip_len field.
The packet was captured in 2012, but its origin is unclear [1].
Presuming it's not a valid packet, remove this test.

[1]: https://github.com/the-tcpdump-group/tcpdump/pull/1329#issuecomment-3356640061

Signed-off-by: Maxim Mikityanskiy <maxim@isovalent.com>