tests: Add coverage for GSO/TSO and BIG TCP
This commit adds coverage that affirms the current behavior of tcpdump
when it parses packets captured in Linux with aggregation offloads
enabled (such as GRO, GSO, TSO and BIG TCP).

New tests include:

1. GRO/GSO/TSO TCP packets. These have ip_len (IPv4) or payload_len
(IPv6) set to a value bigger than MTU, but smaller than 64k (fits the
width of the length field, which is 16 bits).

2. BIG TCP packets, as currently implemented in Linux. These aggregated
packets are larger than 64k. In IPv4, ip_len is just set to 0. In IPv6,
payload_len is also set to 0, and the real length of the payload is
indicated in the Jumbo option of the Hop-by-hop extension header,
immediately following the IPv6 header.

3. A corner case: a small (<64k) IPv4 packet with ip_len set to 0. Turns
out that both tcpdump and wireshark treat such packets as valid,
regardless of the transport protocol, simply printing a message that TSO
was presumed. While my tests show that TSO actually fills in the correct
(non-zero) value into the ip_len field, the test suite already contains
a test called ipv4_tcp_http_xml_tso with a 2030-byte TCP/IPv4 packet
that has ip_len = 0, but it's treated as a valid packet. The packet was
captured in 2012, and the exact circumstances of why it has a zero
ip_len are unclear to me, but if such packets are considered valid by
both tcpdump and wireshark, I'm extending the coverage by adding an ICMP
packet with ip_len = 0, which was manually crafted.

Signed-off-by: Maxim Mikityanskiy <maxim@isovalent.com>